test: one UnityPy stub instead of five
The `_install_unitypy_stub` fixture was pasted into five test modules: three
byte-identical copies (test_decrypt_and_extract{,_more,_unit}.py) and two
subset copies (test_extract_client_data{,_unit}.py) that omit
helpers.ArchiveStorageManager and set_assetbundle_decrypt_key.

The subset copies made the fixture order-sensitive — whichever module pytest
collected first decided what the stub contained — which is why the full copies
carry a `hasattr(..., "helpers")` upgrade guard. tests/unitypy_stub.py always
installs the full stub, so that hazard goes away with the duplication; the
subset consumers only ever touch the subset.

All five modules verified passing standalone as well as in the full suite
(2999 green), so the shared fixture is not quietly relying on collection order
either.
