+++ tests/test_extract_client_data.py
@@ -7,7 +7,6 @@ UnityPy / PIL / client files are touched; all I/O runs through tempfile.
 
 import sys
 import tempfile
-import types
 import unittest
 from pathlib import Path
 from unittest import mock
@@ -15,38 +14,11 @@ from unittest import mock
 SCRIPTS = Path(__file__).resolve().parent.parent / "projects" / "wiki" / "scripts"
 sys.path.insert(0, str(SCRIPTS))
 
+from unitypy_stub import install_unitypy_stub  # noqa: E402  桩单一真相源
+
 
 # --- stub UnityPy + UnityPy.enums so the module imports without the real lib ---
-def _install_unitypy_stub():
-    if "UnityPy" in sys.modules:
-        return
-
-    class _ClassIDType:
-        # The module only references these three names.
-        class _T:
-            def __init__(self, name):
-                self.name = name
-
-            def __eq__(self, other):
-                return isinstance(other, _ClassIDType._T) and other.name == self.name
-
-            def __hash__(self):
-                return hash(self.name)
-
-        TextAsset = _T("TextAsset")
-        MonoBehaviour = _T("MonoBehaviour")
-        Texture2D = _T("Texture2D")
-
-    unitypy = types.ModuleType("UnityPy")
-    unitypy.load = mock.MagicMock(name="UnityPy.load")
-    enums = types.ModuleType("UnityPy.enums")
-    enums.ClassIDType = _ClassIDType
-    unitypy.enums = enums
-    sys.modules["UnityPy"] = unitypy
-    sys.modules["UnityPy.enums"] = enums
-
-
-_install_unitypy_stub()
+install_unitypy_stub()
 
 import extract_client_data as ecd  # noqa: E402
 from UnityPy.enums import ClassIDType  # noqa: E402
