+++ tests/test_extract_client_data_unit.py
@@ -5,7 +5,6 @@ UnityPy/PIL stubbed via sys.modules; all I/O uses tmp_path.
 
 import json
 import sys
-import types
 from pathlib import Path
 from unittest import mock
 
@@ -14,36 +13,8 @@ import pytest
 SCRIPTS = Path(__file__).resolve().parent.parent / "projects" / "wiki" / "scripts"
 sys.path.insert(0, str(SCRIPTS))
 
-
-def _install_unitypy_stub():
-    if "UnityPy" in sys.modules:
-        return
-
-    class _ClassIDType:
-        class _T:
-            def __init__(self, name):
-                self.name = name
-
-            def __eq__(self, other):
-                return isinstance(other, _ClassIDType._T) and other.name == self.name
-
-            def __hash__(self):
-                return hash(self.name)
-
-        TextAsset = _T("TextAsset")
-        MonoBehaviour = _T("MonoBehaviour")
-        Texture2D = _T("Texture2D")
-
-    unitypy = types.ModuleType("UnityPy")
-    unitypy.load = mock.MagicMock(name="UnityPy.load")
-    enums = types.ModuleType("UnityPy.enums")
-    enums.ClassIDType = _ClassIDType
-    unitypy.enums = enums
-    sys.modules["UnityPy"] = unitypy
-    sys.modules["UnityPy.enums"] = enums
-
-
-_install_unitypy_stub()
+from unitypy_stub import install_unitypy_stub  # noqa: E402  桩单一真相源
+install_unitypy_stub()
 
 import extract_client_data as ecd  # noqa: E402
 from UnityPy.enums import ClassIDType  # noqa: E402
